After building the behavior interfaces and the rest of the classes that implement them, it's easier for me to see how this all works. I still have some questions about my Duck class.
Although, generally up to this point, this is the material that I had a fairly decent understanding of. Taking it further into full out design patterns and using techniques such as DIP and LSP are still pretty much like playing a mental game of Twister. I'll get it eventually, I just need time...lots of time.